Additional type changes required for MacOS build (#994)
On MacOS `uint64_t` and `std::size_t` are different types. This PR replaces uses of `size_t` by `uint64_t` where the latter type is more appropriate. It also replaces `long int` by `int64_t`, as those types are also not the same on MacOS.

+++ src/engine/CountAvailablePredicates.cpp
@@ -79,7 +79,7 @@ float CountAvailablePredicates::getMultiplicity([[maybe_unused]] size_t col) {
 }
 
 // _____________________________________________________________________________
-size_t CountAvailablePredicates::getSizeEstimateBeforeLimit() {
+uint64_t CountAvailablePredicates::getSizeEstimateBeforeLimit() {
   if (_subtree.get() != nullptr) {
     // Predicates are only computed for entities in the subtrees result.
